pyproject.toml
[tool.poetry]
name = "rag-improvised-optimized"
version = "0.1.0"
description = "Optimized RAG agent system for Hugging Face."
authors = ["Harshil Awasthi"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
gunicorn = "^22.0.0"
pydantic = "^2.7.1"
google-generativeai = "^0.5.4"
chromadb = "^0.5.0"
sentence-transformers = "^2.7.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
numpy = "^1.26.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

